cps: move routing daemon code to new file cps/rd.c
The code that supports routing daemons used to live in cps/kni.c.
This patch moves this code to the new file cps/rd.c.

This move splits cps/kni.c into two files about even in
the number of lines. The file cps/kni.c had more tahn 2K linas
before this patch.
