+++ cps/kni.c
@@ -16,13 +16,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <arpa/inet.h>
 #include <fcntl.h>
 #include <libkmod.h>
 #include <libmnl/libmnl.h>
 #include <linux/rtnetlink.h>
 #include <net/if.h>
-#include <net/if_arp.h>
 #include <stdbool.h>
 #include <sys/stat.h>
 #include <sys/utsname.h>
@@ -40,46 +38,12 @@
 
 #define KNI_MODULE_NAME ("rte_kni")
 
-/* Defined in the kernel headers, but not included in net/if.h. */
-#define IFF_LOWER_UP (1<<16)
-
 /*
  * According to init_module(2) and delete_module(2), there
  * are no declarations for these functions in header files.
  */
 extern long init_module(void *, unsigned long, const char *);
 extern long delete_module(const char *, unsigned int);
-
-struct route_update {
-	/* Type of route update: RTM_NEWROUTE or RTM_DELROUTE. */
-	int      type;
-
-	/* Address family of route: AF_INET or AF_INET6. */
-	int      family;
-
-	/*
-	 * Whether this update has all the fields and attributes
-	 * necessary to update the LPM table.
-	 */
-	int      valid;
-
-	/* Route origin. See field rtm_protocol of struct rtmsg. */
-	uint8_t  rt_proto;
-
-	/* Route type. See field rtm_type of struct rtmsg. */
-	uint8_t  rt_type;
-
-	/* Output interface index of route. */
-	uint32_t oif_index;
-
-	/* The network prefix (destination) of route. */
-	char ip_px_buf[INET6_ADDRSTRLEN + 4];
-	struct ip_prefix prefix_info;
-
-	/* IP address of route gateway. */
-	char gw_buf[INET6_ADDRSTRLEN];
-	struct ipaddr gw;
-};
 
 int
 kni_change_if(uint16_t port_id, uint8_t if_up)
@@ -400,1129 +364,6 @@ kni_config_link(struct rte_kni *kni)
 close:
 	mnl_socket_close(nl);
 	return ret;
-}
-
-/* The below rd_*() functions handle interactions with the routing daemon. */
-
-void
-rd_event_sock_close(struct cps_config *cps_conf)
-{
-	if (cps_conf->rd_nl != NULL) {
-		mnl_socket_close(cps_conf->rd_nl);
-		cps_conf->rd_nl = NULL;
-	}
-}
-
-int
-rd_event_sock_open(struct cps_config *cps_conf)
-{
-	struct mnl_socket *nl;
-	int ret;
-
-	nl = mnl_socket_open(NETLINK_ROUTE);
-	if (nl == NULL) {
-		CPS_LOG(ERR, "%s: mnl_socket_open: %s\n",
-			__func__, strerror(errno));
-		return -1;
-	}
-	cps_conf->rd_nl = nl;
-
-	/*
-	 * This binds the Netlink socket to port @nl_pid,
-	 * so the routing daemon may interact with Gatekeeper.
-	 */
-	ret = mnl_socket_bind(nl, 0, cps_conf->nl_pid);
-	if (ret < 0) {
-		CPS_LOG(ERR, "%s: mnl_socket_bind: %s\n",
-			__func__, strerror(errno));
-		goto close;
-	}
-
-	return 0;
-
-close:
-	rd_event_sock_close(cps_conf);
-	return ret;
-}
-
-static int
-new_route(struct route_update *update, const struct cps_config *cps_conf)
-{
-	int ret;
-	struct gk_fib *gw_fib = NULL;
-	struct gk_lpm *ltbl = &cps_conf->gk->lpm_tbl;
-
-	if (update->family == AF_INET) {
-		if (update->rt_type != RTN_BLACKHOLE) {
-			ret = lpm_lookup_ipv4(ltbl->lpm,
-				update->gw.ip.v4.s_addr);
-			if (ret < 0) {
-				if (ret == -ENOENT) {
-					CPS_LOG(WARNING,
-						"%s(): there is no route to the gateway %s of the IPv4 route %s sent by routing daemon\n",
-						__func__, update->gw_buf,
-						update->ip_px_buf);
-				}
-				return ret;
-			}
-			gw_fib = &ltbl->fib_tbl[ret];
-		}
-	} else if (likely(update->family == AF_INET6)) {
-		if (update->rt_type != RTN_BLACKHOLE) {
-			ret = lpm_lookup_ipv6(ltbl->lpm6, &update->gw.ip.v6);
-			if (ret < 0) {
-				if (ret == -ENOENT) {
-					CPS_LOG(WARNING,
-						"%s(): there is no route to the gateway %s of the IPv6 route %s sent by routing daemon\n",
-						__func__, update->gw_buf,
-						update->ip_px_buf);
-				}
-				return ret;
-			}
-			gw_fib = &ltbl->fib_tbl6[ret];
-		}
-	} else {
-		CPS_LOG(WARNING,
-			"cps update: unknown address family %d at %s()\n",
-			update->family, __func__);
-		return -EAFNOSUPPORT;
-	}
-
-	if (update->rt_type == RTN_BLACKHOLE) {
-		return add_fib_entry_numerical(&update->prefix_info, NULL,
-			NULL, 0, GK_DROP, update->rt_proto, cps_conf->gk);
-	}
-
-	RTE_VERIFY(gw_fib != NULL);
-
-	if (gw_fib->action == GK_FWD_NEIGHBOR_FRONT_NET) {
-		if (update->oif_index != 0) {
-			if (update->oif_index != cps_conf->front_kni_index) {
-				CPS_LOG(WARNING,
-					"The output KNI interface for prefix %s is not the front interface while the gateway %s is a neighbor of the front network\n",
-					update->ip_px_buf, update->gw_buf);
-				return -EINVAL;
-			}
-		}
-
-		return add_fib_entry_numerical(&update->prefix_info, NULL,
-			&update->gw, 1, GK_FWD_GATEWAY_FRONT_NET,
-			update->rt_proto, cps_conf->gk);
-	}
-
-	if (gw_fib->action == GK_FWD_NEIGHBOR_BACK_NET) {
-		if (update->oif_index != 0) {
-			if (update->oif_index != cps_conf->back_kni_index) {
-				CPS_LOG(WARNING,
-					"The output KNI interface for prefix %s is not the back interface while the gateway %s is a neighbor of the back network\n",
-					update->ip_px_buf, update->gw_buf);
-				return -EINVAL;
-			}
-		}
-
-		return add_fib_entry_numerical(&update->prefix_info, NULL,
-			&update->gw, 1, GK_FWD_GATEWAY_BACK_NET,
-			update->rt_proto, cps_conf->gk);
-	}
-
-	return -EINVAL;
-}
-
-/*
- * @addr_buf must be at least INET6_ADDRSTRLEN bytes long.
- * @addr can be NULL.
- */
-static int
-__convert_ip_attr(int family, struct nlattr *tb[], enum rtattr_type_t attr_type,
-	const char *attr_name, struct ipaddr *addr, char *addr_buf)
-{
-	if (family == AF_INET) {
-		const struct in_addr *addr4 =
-			mnl_attr_get_payload(tb[attr_type]);
-		if (addr) {
-			addr->proto = RTE_ETHER_TYPE_IPV4;
-			addr->ip.v4 = *addr4;
-		}
-
-		if (unlikely(inet_ntop(AF_INET, addr4, addr_buf,
-				INET_ADDRSTRLEN) == NULL)) {
-			int saved_errno = errno;
-			CPS_LOG(ERR,
-				"%s(%s): failed to convert an IPv4 address: %s\n",
-				__func__, attr_name, strerror(errno));
-			return -saved_errno;
-		}
-
-		return 0;
-	}
-
-	if (likely(family == AF_INET6)) {
-		const struct in6_addr *addr6 =
-			mnl_attr_get_payload(tb[attr_type]);
-		if (addr) {
-			addr->proto = RTE_ETHER_TYPE_IPV6;
-			addr->ip.v6 = *addr6;
-		}
-
-		if (unlikely(inet_ntop(AF_INET6, addr6, addr_buf,
-				INET6_ADDRSTRLEN) == NULL)) {
-			int saved_errno = errno;
-			CPS_LOG(ERR,
-				"%s(%s): failed to convert an IPv6 address: %s\n",
-				__func__, attr_name, strerror(errno));
-			return -saved_errno;
-		}
-
-		return 0;
-	}
-
-	CPS_LOG(WARNING, "%s(%s): unknown address family %d\n",
-			__func__, attr_name, family);
-	return -EAFNOSUPPORT;
-}
-
-#define convert_ip_attr(family, tb, attr_type, addr, addr_buf)	\
-	__convert_ip_attr(family, tb, attr_type, #attr_type, addr, addr_buf)
-
-static int
-attr_get(struct route_update *update, int family, struct nlattr *tb[])
-{
-	char addr_buf[INET6_ADDRSTRLEN];
-	int ret;
-	bool dst_present = false;
-	bool gw_present = false;
-	bool oif_present = false;
-
-	if (tb[RTA_MULTIPATH]) {
-		/*
-		 * XXX #75 This is the attribute used to implement ECMP.
-		 * We should more closely parse this attribute and
-		 * return the appropriate information through
-		 * @update to Grantor, if we're running Grantor.
-		 *
-		 * Example usage:
-		 *
-		 * struct rtnexthop *rt =
-		 *	mnl_attr_get_payload(tb[RTA_MULTIPATH]);
-		 */
-		CPS_LOG(WARNING,
-			"cps update: the rtnetlink command has information (RTA_MULTIPATH) that we don't need or don't honor\n");
-	}
-
-	if (tb[RTA_DST]) {
-		ret = convert_ip_attr(family, tb, RTA_DST,
-			&update->prefix_info.addr, addr_buf);
-		if (ret < 0)
-			return ret;
-
-		/* Fill in prefix string. */
-		ret = snprintf(update->ip_px_buf, sizeof(update->ip_px_buf),
-				"%s/%hhu", addr_buf, update->prefix_info.len);
-		RTE_VERIFY(ret > 0 && ret < (int)sizeof(update->ip_px_buf));
-		update->prefix_info.str = update->ip_px_buf;
-
-		CPS_LOG(DEBUG, "cps update: dst: %s\n", update->ip_px_buf);
-		dst_present = true;
-	}
-
-	if (tb[RTA_SRC]) {
-		ret = convert_ip_attr(family, tb, RTA_SRC, NULL, addr_buf);
-		if (ret < 0)
-			return ret;
-
-		CPS_LOG(WARNING,
-			"cps update: the rtnetlink command has information (RTA_SRC with IP address %s) that we don't need or don't honor\n",
-			addr_buf);
-	}
-
-	if (tb[RTA_OIF]) {
-		update->oif_index = mnl_attr_get_u32(tb[RTA_OIF]);
-		CPS_LOG(DEBUG, "cps update: oif=%u\n", update->oif_index);
-		oif_present = true;
-	}
-
-	if (tb[RTA_FLOW]) {
-		CPS_LOG(WARNING,
-			"cps update: the rtnetlink command has information (RTA_FLOW with flow=%u) that we don't need or don't honor\n",
-			mnl_attr_get_u32(tb[RTA_FLOW]));
-	}
-
-	if (tb[RTA_PREFSRC]) {
-		ret = convert_ip_attr(family, tb, RTA_PREFSRC, NULL, addr_buf);
-		if (ret < 0)
-			return ret;
-
-		CPS_LOG(WARNING,
-			"cps update: the rtnetlink command has information (RTA_PREFSRC with IP address %s) that we don't need or don't honor\n",
-			addr_buf);
-	}
-
-	if (tb[RTA_GATEWAY]) {
-		ret = convert_ip_attr(family, tb, RTA_GATEWAY,
-			&update->gw, update->gw_buf);
-		if (ret < 0)
-			return ret;
-
-		CPS_LOG(DEBUG, "cps update: gw: %s\n", update->gw_buf);
-		gw_present = true;
-	}
-
-	if (tb[RTA_PRIORITY]) {
-		CPS_LOG(WARNING,
-			"cps update: the rtnetlink command has information (RTA_PRIORITY with prio=%u) that we don't need or don't honor\n",
-			mnl_attr_get_u32(tb[RTA_PRIORITY]));
-	}
-
-	update->valid = dst_present &&
-		(update->type == RTM_DELROUTE ||
-		(update->type == RTM_NEWROUTE && gw_present && oif_present) ||
-		(update->type == RTM_NEWROUTE &&
-			update->rt_type == RTN_BLACKHOLE));
-	return 0;
-}
-
-static int
-data_ipv4_attr_cb(const struct nlattr *attr, void *data)
-{
-	const struct nlattr **tb = data;
-	int type = mnl_attr_get_type(attr);
-
-	/* Skip unsupported attribute in user-space. */
-	if (mnl_attr_type_valid(attr, RTA_MAX) < 0)
-		return MNL_CB_OK;
-
-	switch (type) {
-	case RTA_MULTIPATH:
-		if (mnl_attr_validate(attr, MNL_TYPE_NESTED) < 0)
-			return MNL_CB_ERROR;
-		break;
-	case RTA_TABLE:
-	case RTA_DST:
-	case RTA_SRC:
-	case RTA_OIF:
-	case RTA_FLOW:
-	case RTA_PREFSRC:
-	case RTA_GATEWAY:
-	case RTA_PRIORITY:
-		if (mnl_attr_validate(attr, MNL_TYPE_U32) < 0)
-			return MNL_CB_ERROR;
-		break;
-	default:
-		/* Skip attributes we don't know about. */
-		return MNL_CB_OK;
-	}
-	tb[type] = attr;
-	return MNL_CB_OK;
-}
-
-static int
-data_ipv6_attr_cb(const struct nlattr *attr, void *data)
-{
-	const struct nlattr **tb = data;
-	int type = mnl_attr_get_type(attr);
-
-	/* Skip unsupported attribute in user-space. */
-	if (mnl_attr_type_valid(attr, RTA_MAX) < 0)
-		return MNL_CB_OK;
-
-	switch (type) {
-	case RTA_MULTIPATH:
-		if (mnl_attr_validate(attr, MNL_TYPE_NESTED) < 0)
-			return MNL_CB_ERROR;
-		break;
-	case RTA_TABLE:
-	case RTA_OIF:
-	case RTA_FLOW:
-	case RTA_PRIORITY:
-		if (mnl_attr_validate(attr, MNL_TYPE_U32) < 0)
-			return MNL_CB_ERROR;
-		break;
-	case RTA_DST:
-	case RTA_SRC:
-	case RTA_PREFSRC:
-	case RTA_GATEWAY:
-		if (mnl_attr_validate2(attr, MNL_TYPE_BINARY,
-				sizeof(struct in6_addr)) < 0)
-			return MNL_CB_ERROR;
-		break;
-	default:
-		/* Skip attributes we don't know about. */
-		return MNL_CB_OK;
-	}
-	tb[type] = attr;
-	return MNL_CB_OK;
-}
-
-static void
-rd_send_err(const struct nlmsghdr *req, struct cps_config *cps_conf, int err)
-{
-	struct nlmsghdr *rep;
-	struct nlmsgerr *errmsg;
-	char buf[MNL_SOCKET_BUFFER_SIZE];
-	struct sockaddr_nl rd_sa;
-	unsigned int payload_len;
-	unsigned int errmsg_len;
-
-	memset(&rd_sa, 0, sizeof(rd_sa));
-	rd_sa.nl_family = AF_NETLINK;
-	rd_sa.nl_pid = req->nlmsg_pid;
-
-	rep = mnl_nlmsg_put_header(buf);
-	rep->nlmsg_type = NLMSG_ERROR;
-	rep->nlmsg_flags = 0;
-	rep->nlmsg_seq = req->nlmsg_seq;
-	rep->nlmsg_pid = cps_conf->nl_pid;
-
-	/*
-	 * For acknowledgements, just send the struct nlmsgerr.
-	 * For errors, send the struct nlmsgerr and the payload.
-	 */
-	payload_len = 0;
-	errmsg_len = sizeof(*errmsg);
-	if (err) {
-		payload_len += mnl_nlmsg_get_payload_len(req);
-		errmsg_len += payload_len;
-	}
-
-	errmsg = mnl_nlmsg_put_extra_header(rep, errmsg_len);
-	errmsg->error = err;
-	memcpy(&errmsg->msg, req, sizeof(errmsg->msg) + payload_len);
-
-	if (sendto(mnl_socket_get_fd(cps_conf->rd_nl),
-			rep, sizeof(*rep) + errmsg_len, 0,
-			(struct sockaddr *)&rd_sa, sizeof(rd_sa)) < 0) {
-		CPS_LOG(ERR, "sendto: cannot send NLMSG_ERROR to daemon (pid=%u seq=%u): %s\n",
-			req->nlmsg_pid, req->nlmsg_seq,
-			strerror(errno));
-	}
-}
-
-static void
-rd_fill_getroute_reply(const struct cps_config *cps_conf,
-	struct nlmsghdr *reply, struct gk_fib *fib, int family, uint32_t seq,
-	uint8_t prefix_len, struct ipaddr **gw_addr)
-{
-	struct rtmsg *rm;
-
-	reply->nlmsg_type = RTM_NEWROUTE;
-	reply->nlmsg_flags = NLM_F_MULTI;
-	reply->nlmsg_seq = seq;
-	reply->nlmsg_pid = cps_conf->nl_pid;
-
-	rm = mnl_nlmsg_put_extra_header(reply, sizeof(*rm));
-	rm->rtm_family = family;
-	rm->rtm_dst_len = prefix_len;
-	rm->rtm_src_len = 0;
-	rm->rtm_tos = 0;
-	rm->rtm_table = RT_TABLE_MAIN;
-	rm->rtm_scope = RT_SCOPE_UNIVERSE;
-	rm->rtm_flags = 0;
-
-	switch (fib->action) {
-	case GK_FWD_GRANTOR:
-		mnl_attr_put_u32(reply, RTA_OIF, cps_conf->back_kni_index);
-		rm->rtm_protocol = RTPROT_STATIC;
-		rm->rtm_type = RTN_UNICAST;
-		/*
-		 * Gateway will be filled in by the caller, since Grantor
-		 * entries can have multiple corresponding Grantors, each
-		 * with their own gateway.
-		 */
-		break;
-	case GK_FWD_GATEWAY_FRONT_NET:
-		mnl_attr_put_u32(reply, RTA_OIF, cps_conf->front_kni_index);
-		rm->rtm_protocol = fib->u.gateway.rt_proto;
-		rm->rtm_type = RTN_UNICAST;
-		*gw_addr = &fib->u.gateway.eth_cache->ip_addr;
-		break;
-	case GK_FWD_GATEWAY_BACK_NET:
-		mnl_attr_put_u32(reply, RTA_OIF, cps_conf->back_kni_index);
-		rm->rtm_protocol = fib->u.gateway.rt_proto;
-		rm->rtm_type = RTN_UNICAST;
-		*gw_addr = &fib->u.gateway.eth_cache->ip_addr;
-		break;
-	case GK_FWD_NEIGHBOR_FRONT_NET:
-		mnl_attr_put_u32(reply, RTA_OIF, cps_conf->front_kni_index);
-		rm->rtm_protocol = RTPROT_STATIC;
-		rm->rtm_type = RTN_UNICAST;
-		*gw_addr = NULL;
-		break;
-	case GK_FWD_NEIGHBOR_BACK_NET:
-		mnl_attr_put_u32(reply, RTA_OIF, cps_conf->back_kni_index);
-		rm->rtm_protocol = RTPROT_STATIC;
-		rm->rtm_type = RTN_UNICAST;
-		*gw_addr = NULL;
-		break;
-	case GK_DROP:
-		rm->rtm_protocol = fib->u.drop.rt_proto;
-		rm->rtm_type = RTN_BLACKHOLE;
-		*gw_addr = NULL;
-		break;
-	default:
-		rte_panic("Invalid FIB action (%u) in FIB while being processed by CPS block in %s\n",
-			fib->action, __func__);
-		return;
-	}
-}
-
-static int
-rd_send_batch(const struct cps_config *cps_conf, struct mnl_nlmsg_batch *batch,
-	const char *daemon, uint32_t seq, uint32_t pid, int done)
-{
-	/* Address of routing daemon. */
-	struct sockaddr_nl rd_sa;
-	int ret = 0;
-
-	if (done) {
-		struct nlmsghdr *done =
-			mnl_nlmsg_put_header(mnl_nlmsg_batch_current(batch));
-		done->nlmsg_type = NLMSG_DONE;
-		done->nlmsg_flags = NLM_F_MULTI;
-		done->nlmsg_seq = seq;
-		done->nlmsg_pid = cps_conf->nl_pid;
-		if (!mnl_nlmsg_batch_next(batch)) {
-			/* Send the *full* batch without the DONE message. */
-			ret = rd_send_batch(cps_conf, batch,
-				daemon, seq, pid, false);
-			if (ret < 0)
-				return ret;
-			/* Go on to send the DONE message. */
-		}
-	}
-
-	memset(&rd_sa, 0, sizeof(rd_sa));
-	rd_sa.nl_family = AF_NETLINK;
-	rd_sa.nl_pid = pid;
-
-	if (sendto(mnl_socket_get_fd(cps_conf->rd_nl),
-			mnl_nlmsg_batch_head(batch),
-			mnl_nlmsg_batch_size(batch), 0,
-			(struct sockaddr *)&rd_sa, sizeof(rd_sa)) < 0) {
-		ret = -errno;
-		CPS_LOG(ERR,
-			"sendto: cannot dump route batch to %s daemon (pid=%u seq=%u): %s\n",
-			daemon, pid, seq, strerror(errno));
-	}
-
-	mnl_nlmsg_batch_reset(batch);
-	return ret;
-}
-
-static int
-rd_getroute_ipv4_locked(const struct cps_config *cps_conf, struct gk_lpm *ltbl,
-	struct mnl_nlmsg_batch *batch, const struct nlmsghdr *req)
-{
-	struct rte_lpm_iterator_state state;
-	const struct rte_lpm_rule *re4;
-	int index;
-
-	int ret = rte_lpm_iterator_state_init(ltbl->lpm, 0, 0, &state);
-	if (ret < 0) {
-		CPS_LOG(ERR, "Failed to initialize the IPv4 LPM rule iterator state in %s\n",
-			__func__);
-		return ret;
-	}
-
-	index = rte_lpm_rule_iterate(&state, &re4);
-	while (index >= 0) {
-		struct gk_fib *fib = &ltbl->fib_tbl[re4->next_hop];
-		struct ipaddr *gw_addr;
-		struct nlmsghdr *reply =
-			mnl_nlmsg_put_header(mnl_nlmsg_batch_current(batch));
-
-		rd_fill_getroute_reply(cps_conf, reply, fib,
-			AF_INET, req->nlmsg_seq, state.depth, &gw_addr);
-
-		/* Add address. */
-		mnl_attr_put_u32(reply, RTA_DST, htonl(re4->ip));
-
-		if (fib->action == GK_FWD_GRANTOR) {
-			unsigned int i;
-			for (i = 0; i < fib->u.grantor.set->num_entries; i++) {
-				gw_addr = &fib->u.grantor.set->entries[i]
-					.eth_cache->ip_addr;
-				mnl_attr_put_u32(reply, RTA_GATEWAY,
-					gw_addr->ip.v4.s_addr);
-			}
-		} else if (gw_addr != NULL) {
-			/* Only report gateway for main routes. */
-			mnl_attr_put_u32(reply, RTA_GATEWAY,
-				gw_addr->ip.v4.s_addr);
-		}
-
-		if (!mnl_nlmsg_batch_next(batch)) {
-			ret = rd_send_batch(cps_conf, batch, "IPv4",
-				req->nlmsg_seq, req->nlmsg_pid, false);
-			if (ret < 0)
-				return ret;
-		}
-
-		index = rte_lpm_rule_iterate(&state, &re4);
-	}
-
-	return 0;
-}
-
-static int
-rd_getroute_ipv6_locked(const struct cps_config *cps_conf, struct gk_lpm *ltbl,
-	struct mnl_nlmsg_batch *batch, const struct nlmsghdr *req)
-{
-	struct rte_lpm6_iterator_state state6;
-	struct rte_lpm6_rule re6;
-	int index;
-
-	int ret = rte_lpm6_iterator_state_init(ltbl->lpm6, 0, 0, &state6);
-	if (ret < 0) {
-		CPS_LOG(ERR, "Failed to initialize the IPv6 LPM rule iterator state in %s\n",
-			__func__);
-		return ret;
-	}
-
-	index = rte_lpm6_rule_iterate(&state6, &re6);
-	while (index >= 0) {
-		struct gk_fib *fib = &ltbl->fib_tbl6[re6.next_hop];
-		struct ipaddr *gw_addr;
-		struct nlmsghdr *reply =
-			mnl_nlmsg_put_header(mnl_nlmsg_batch_current(batch));
-
-		rd_fill_getroute_reply(cps_conf, reply, fib,
-			AF_INET6, req->nlmsg_seq, state6.depth, &gw_addr);
-
-		/* Add address. */
-		mnl_attr_put(reply, RTA_DST,
-			sizeof(struct in6_addr), re6.ip);
-
-		if (fib->action == GK_FWD_GRANTOR) {
-			unsigned int i;
-			for (i = 0; i < fib->u.grantor.set->num_entries; i++) {
-				gw_addr = &fib->u.grantor.set->entries[i]
-					.eth_cache->ip_addr;
-				mnl_attr_put(reply, RTA_GATEWAY,
-					sizeof(struct in6_addr),
-					&gw_addr->ip.v6);
-			}
-		} else if (gw_addr != NULL) {
-			/* Only report gateway for main routes. */
-			mnl_attr_put(reply, RTA_GATEWAY,
-				sizeof(struct in6_addr), &gw_addr->ip.v6);
-		}
-
-		if (!mnl_nlmsg_batch_next(batch)) {
-			ret = rd_send_batch(cps_conf, batch, "IPv6",
-				req->nlmsg_seq, req->nlmsg_pid, false);
-			if (ret < 0)
-				return ret;
-		}
-
-		index = rte_lpm6_rule_iterate(&state6, &re6);
-	}
-
-	return 0;
-}
-
-static int
-rd_getroute(const struct nlmsghdr *req, const struct cps_config *cps_conf,
-	int *err)
-{
-	/*
-	 * Buffer length set according to libmnl documentation:
-	 * the buffer that you have to use to store the batch must be
-	 * double of MNL_SOCKET_BUFFER_SIZE to ensure that the last
-	 * message (message N+1) that did not fit into the batch is
-	 * written inside valid memory boundaries.
-	 */
-	char buf[2 * MNL_SOCKET_BUFFER_SIZE];
-	struct mnl_nlmsg_batch *batch;
-	struct gk_lpm *ltbl = &cps_conf->gk->lpm_tbl;
-	const char *family_str;
-	int family;
-
-	if (mnl_nlmsg_get_payload_len(req) < sizeof(struct rtgenmsg)) {
-		CPS_LOG(ERR, "Not enough room in CPS GETROUTE message from routing daemon in %s\n",
-			__func__);
-		*err = -EINVAL;
-		goto out;
-	}
-
-	family = ((struct rtgenmsg *)mnl_nlmsg_get_payload(req))->rtgen_family;
-
-	switch (family) {
-	case AF_INET:
-		family_str = "IPv4";
-		break;
-	case AF_INET6:
-		family_str = "IPv6";
-		break;
-	case AF_UNSPEC:
-		family_str = "IPV4/IPv6";
-		break;
-	case AF_MPLS:
-		family_str = "MPLS";
-		break;
-	default:
-		CPS_LOG(ERR, "Unsupported address family type (%d) in %s\n",
-			family, __func__);
-		*err = -EAFNOSUPPORT;
-		goto out;
-	}
-
-	batch = mnl_nlmsg_batch_start(buf, MNL_SOCKET_BUFFER_SIZE);
-	if (batch == NULL) {
-		CPS_LOG(ERR, "Failed to allocate a batch for a GETROUTE reply\n");
-		*err = -ENOMEM;
-		goto out;
-	}
-
-	if (family == AF_INET || family == AF_UNSPEC) {
-		if (!ipv4_configured(cps_conf->net)) {
-			if (family == AF_UNSPEC)
-				goto ipv6;
-			else {
-				*err = -EAFNOSUPPORT;
-				goto free_batch;
-			}
-		}
-
-		rte_spinlock_lock_tm(&ltbl->lock);
-		*err = rd_getroute_ipv4_locked(cps_conf, ltbl, batch, req);
-		rte_spinlock_unlock_tm(&ltbl->lock);
-		if (*err < 0)
-			goto free_batch;
-	}
-ipv6:
-	if (family == AF_INET6 || family == AF_UNSPEC) {
-		if (!ipv6_configured(cps_conf->net)) {
-			if (family == AF_UNSPEC)
-				goto send;
-			else {
-				*err = -EAFNOSUPPORT;
-				goto free_batch;
-			}
-		}
-
-		rte_spinlock_lock_tm(&ltbl->lock);
-		*err = rd_getroute_ipv6_locked(cps_conf, ltbl, batch, req);
-		rte_spinlock_unlock_tm(&ltbl->lock);
-		if (*err < 0)
-			goto free_batch;
-	}
-send:
-	/* In the case of no entries, the only message sent is NLMSG_DONE. */
-	*err = rd_send_batch(cps_conf, batch, family_str,
-		req->nlmsg_seq, req->nlmsg_pid, true);
-
-free_batch:
-	mnl_nlmsg_batch_stop(batch);
-out:
-	return MNL_CB_OK;
-}
-
-static void
-rd_fill_getlink_reply(const struct cps_config *cps_conf,
-	struct mnl_nlmsg_batch *batch,
-	const char *kni_name, unsigned int kni_index, unsigned int kni_mtu,
-	uint32_t seq)
-{
-	struct nlmsghdr *reply;
-	struct ifinfomsg *ifim;
-
-	reply = mnl_nlmsg_put_header(mnl_nlmsg_batch_current(batch));
-	reply->nlmsg_type = RTM_NEWLINK;
-	reply->nlmsg_flags = NLM_F_MULTI;
-	reply->nlmsg_seq = seq;
-	reply->nlmsg_pid = cps_conf->nl_pid;
-
-	ifim = mnl_nlmsg_put_extra_header(reply, sizeof(*ifim));
-	ifim->ifi_family = AF_UNSPEC;
-	ifim->ifi_type = ARPHRD_ETHER;
-	ifim->ifi_index = kni_index;
-	ifim->ifi_flags = IFF_UP|IFF_LOWER_UP;
-	ifim->ifi_change = 0xFFFFFFFF;
-
-	mnl_attr_put_strz(reply, IFLA_IFNAME, kni_name);
-	mnl_attr_put_u32(reply, IFLA_MTU, kni_mtu);
-}
-
-static int
-rd_getlink(const struct nlmsghdr *req, const struct cps_config *cps_conf,
-	int *err)
-{
-	char buf[2 * MNL_SOCKET_BUFFER_SIZE];
-	struct mnl_nlmsg_batch *batch;
-
-	batch = mnl_nlmsg_batch_start(buf, MNL_SOCKET_BUFFER_SIZE);
-	if (batch == NULL) {
-		CPS_LOG(ERR, "Failed to allocate a batch for a GETLINK reply\n");
-		*err = -ENOMEM;
-		goto out;
-	}
-
-	rd_fill_getlink_reply(cps_conf, batch,
-		rte_kni_get_name(cps_conf->front_kni),
-		cps_conf->front_kni_index, cps_conf->net->front.mtu,
-		req->nlmsg_seq);
-	if (!mnl_nlmsg_batch_next(batch)) {
-		/* Send whatever was in the batch, if anything. */
-		*err = rd_send_batch(cps_conf, batch, "LINK",
-			req->nlmsg_seq, req->nlmsg_pid, false);
-		if (*err < 0)
-			goto free_batch;
-	}
-
-	if (cps_conf->net->back_iface_enabled) {
-		rd_fill_getlink_reply(cps_conf, batch,
-			rte_kni_get_name(cps_conf->back_kni),
-			cps_conf->back_kni_index,
-			cps_conf->net->back.mtu, req->nlmsg_seq);
-		if (!mnl_nlmsg_batch_next(batch)) {
-			*err = rd_send_batch(cps_conf, batch, "LINK",
-				req->nlmsg_seq, req->nlmsg_pid, false);
-			if (*err < 0)
-				goto free_batch;
-		}
-	}
-
-	*err = rd_send_batch(cps_conf, batch, "LINK",
-		req->nlmsg_seq, req->nlmsg_pid, true);
-
-free_batch:
-	mnl_nlmsg_batch_stop(batch);
-out:
-	return MNL_CB_OK;
-}
-
-static int
-rd_modroute(const struct nlmsghdr *req, const struct cps_config *cps_conf,
-	int *err)
-{
-	struct nlattr *tb[__RTA_MAX] = {};
-	struct rtmsg *rm = mnl_nlmsg_get_payload(req);
-	struct route_update update;
-
-	if (unlikely(cps_conf->gk == NULL)) {
-		/*
-		 * Grantor only runs CPS for ECMP support and
-		 * shouldn't be receiving route updates.
-		 */
-		CPS_LOG(WARNING,
-			"The system is running as Grantor, and there shouldn't be any rtnetlink message processed under this configuration while receiving route update messages\n");
-		*err = -EOPNOTSUPP;
-		goto out;
-	}
-
-	CPS_LOG(DEBUG, "cps update: [%s] family=%u dst_len=%u src_len=%u tos=%u table=%u protocol=%u scope=%u type=%u flags=%x\n",
-		req->nlmsg_type == RTM_NEWROUTE ? "NEW" : "DEL",
-		rm->rtm_family, rm->rtm_dst_len, rm->rtm_src_len,
-		rm->rtm_tos, rm->rtm_table, rm->rtm_protocol,
-		rm->rtm_scope, rm->rtm_type, rm->rtm_flags);
-
-	memset(&update, 0, sizeof(update));
-	update.valid = false;
-	update.type = req->nlmsg_type;
-	update.family = rm->rtm_family;
-
-	/* Destination prefix length, e.g., 24 or 32 for IPv4. */
-	update.prefix_info.len = rm->rtm_dst_len;
-
-	/* Default to an invalid index number. */
-	update.oif_index = 0;
-
-	/* Route origin (routing daemon). */
-	update.rt_proto = rm->rtm_protocol;
-
-	/* Route type. */
-	update.rt_type = rm->rtm_type;
-
-	switch (rm->rtm_family) {
-	case AF_INET:
-		if (!ipv4_configured(cps_conf->net)) {
-			*err = -EAFNOSUPPORT;
-			goto out;
-		}
-		mnl_attr_parse(req, sizeof(*rm), data_ipv4_attr_cb, tb);
-		*err = attr_get(&update, rm->rtm_family, tb);
-		if (*err)
-			goto out;
-		break;
-	case AF_INET6:
-		if (!ipv6_configured(cps_conf->net)) {
-			*err = -EAFNOSUPPORT;
-			goto out;
-		}
-		mnl_attr_parse(req, sizeof(*rm), data_ipv6_attr_cb, tb);
-		*err = attr_get(&update, rm->rtm_family, tb);
-		if (*err)
-			goto out;
-		break;
-	default:
-		CPS_LOG(NOTICE, "Unrecognized family in netlink event: %u\n",
-			rm->rtm_family);
-		*err = -EAFNOSUPPORT;
-		goto out;
-	}
-
-	if (update.valid) {
-		if (update.type == RTM_NEWROUTE) {
-			*err = new_route(&update, cps_conf);
-		} else if (likely(update.type == RTM_DELROUTE)) {
-			*err = del_fib_entry_numerical(&update.prefix_info,
-				cps_conf->gk);
-		} else {
-			CPS_LOG(WARNING, "Receiving an unexpected update rule with type = %d\n",
-				update.type);
-			*err = -EOPNOTSUPP;
-		}
-	} else
-		*err = -EINVAL;
-
-out:
-	return MNL_CB_OK;
-}
-
-static void
-rd_fill_getaddr_reply(const struct cps_config *cps_conf,
-	struct mnl_nlmsg_batch *batch, struct gatekeeper_if *iface,
-	uint8_t family, unsigned int kni_index, uint32_t seq)
-{
-	struct nlmsghdr *reply;
-	struct ifaddrmsg *ifam;
-
-	reply = mnl_nlmsg_put_header(mnl_nlmsg_batch_current(batch));
-	reply->nlmsg_type = RTM_NEWADDR;
-	reply->nlmsg_flags = NLM_F_MULTI;
-	reply->nlmsg_seq = seq;
-	reply->nlmsg_pid = cps_conf->nl_pid;
-
-	ifam = mnl_nlmsg_put_extra_header(reply, sizeof(*ifam));
-	ifam->ifa_family = family;
-	ifam->ifa_flags = IFA_F_PERMANENT;
-	ifam->ifa_scope = RT_SCOPE_UNIVERSE;
-	ifam->ifa_index = kni_index;
-
-	/*
-	 * The exact meaning of IFA_LOCAL and IFA_ADDRESS depend
-	 * on the address family being used and the device type.
-	 * For broadcast devices (like the interfaces we use),
-	 * for IPv4 we specify both and they are used interchangeably.
-	 * For IPv6, only IFA_ADDRESS needs to be set.
-	 */
-
-	if (family == AF_INET) {
-		mnl_attr_put_u32(reply, IFA_LOCAL, iface->ip4_addr.s_addr);
-		mnl_attr_put_u32(reply, IFA_ADDRESS, iface->ip4_addr.s_addr);
-		ifam->ifa_prefixlen = iface->ip4_addr_plen;
-	} else if (likely(family == AF_INET6)) {
-		mnl_attr_put(reply, IFA_ADDRESS,
-			sizeof(iface->ip6_addr), &iface->ip6_addr);
-		ifam->ifa_prefixlen = iface->ip6_addr_plen;
-	} else {
-		rte_panic("Invalid address family (%hhu) in request while being processed by CPS block in %s\n",
-			family, __func__);
-	}
-}
-
-static int
-rd_getaddr_iface(const struct cps_config *cps_conf,
-	struct mnl_nlmsg_batch *batch, struct gatekeeper_if *iface,
-	uint8_t family, unsigned int kni_index, uint32_t seq, uint32_t pid)
-{
-	int ret = 0;
-
-	if ((family == AF_INET || family == AF_UNSPEC)
-			&& ipv4_if_configured(iface)) {
-		rd_fill_getaddr_reply(cps_conf, batch, iface,
-			AF_INET, kni_index, seq);
-		if (!mnl_nlmsg_batch_next(batch)) {
-			/* Send whatever was in the batch, if anything. */
-			ret = rd_send_batch(cps_conf, batch, "IPv4",
-				seq, pid, false);
-			if (ret < 0)
-				return ret;
-		}
-	}
-
-	if ((family == AF_INET6 || family == AF_UNSPEC)
-			&& ipv6_if_configured(iface)) {
-		rd_fill_getaddr_reply(cps_conf, batch, iface,
-			AF_INET6, kni_index, seq);
-		if (!mnl_nlmsg_batch_next(batch)) {
-			/* Send whatever was in the batch, if anything. */
-			ret = rd_send_batch(cps_conf, batch, "IPv6",
-				seq, pid, false);
-			if (ret < 0)
-				return ret;
-		}
-	}
-
-	return ret;
-}
-
-static int
-rd_getaddr(const struct nlmsghdr *req, const struct cps_config *cps_conf,
-	int *err)
-{
-	char buf[2 * MNL_SOCKET_BUFFER_SIZE];
-	struct mnl_nlmsg_batch *batch;
-	struct net_config *net_conf = cps_conf->net;
-	int family;
-	const char *family_str;
-
-	if (mnl_nlmsg_get_payload_len(req) < sizeof(struct rtgenmsg)) {
-		CPS_LOG(ERR, "Not enough room in CPS GETADDR message from routing daemon in %s\n",
-			__func__);
-		*err = -EINVAL;
-		goto out;
-	}
-
-	family = ((struct rtgenmsg *)mnl_nlmsg_get_payload(req))->rtgen_family;
-
-	switch (family) {
-	case AF_INET:
-		family_str = "IPv4";
-		break;
-	case AF_INET6:
-		family_str = "IPv6";
-		break;
-	case AF_UNSPEC:
-		family_str = "IPV4/IPv6";
-		break;
-	default:
-		CPS_LOG(ERR, "Unsupported address family type (%d) in %s\n",
-			family, __func__);
-		*err = -EAFNOSUPPORT;
-		goto out;
-	}
-
-	batch = mnl_nlmsg_batch_start(buf, MNL_SOCKET_BUFFER_SIZE);
-	if (batch == NULL) {
-		CPS_LOG(ERR, "Failed to allocate a batch for a GETADDR reply\n");
-		*err = -ENOMEM;
-		goto out;
-	}
-
-	*err = rd_getaddr_iface(cps_conf, batch, &net_conf->front, family,
-		cps_conf->front_kni_index, req->nlmsg_seq, req->nlmsg_pid);
-	if (*err < 0)
-		goto free_batch;
-
-	if (net_conf->back_iface_enabled) {
-		*err = rd_getaddr_iface(cps_conf, batch, &net_conf->back,
-			family, cps_conf->back_kni_index,
-			req->nlmsg_seq, req->nlmsg_pid);
-		if (*err < 0)
-			goto free_batch;
-	}
-
-	*err = rd_send_batch(cps_conf, batch, family_str,
-		req->nlmsg_seq, req->nlmsg_pid, true);
-
-free_batch:
-	mnl_nlmsg_batch_stop(batch);
-out:
-	return MNL_CB_OK;
-}
-
-static int
-rd_cb(const struct nlmsghdr *req, void *arg)
-{
-	struct cps_config *cps_conf = arg;
-	int ret = MNL_CB_OK;
-	int err;
-
-	/* Only requests should be received here. */
-	if (!(req->nlmsg_flags & NLM_F_REQUEST)) {
-		err = -EINVAL;
-		goto out;
-	}
-
-	switch (req->nlmsg_type) {
-	case RTM_NEWROUTE:
-		/* FALLTHROUGH */
-	case RTM_DELROUTE:
-		ret = rd_modroute(req, cps_conf, &err);
-		break;
-	case RTM_GETROUTE:
-		ret = rd_getroute(req, cps_conf, &err);
-		break;
-	case RTM_GETLINK:
-		ret = rd_getlink(req, cps_conf, &err);
-		break;
-	case RTM_GETADDR:
-		ret = rd_getaddr(req, cps_conf, &err);
-		break;
-	default:
-		CPS_LOG(NOTICE, "Unrecognized netlink message type: %u\n",
-			req->nlmsg_type);
-		err = -EOPNOTSUPP;
-		break;
-	}
-out:
-	if ((req->nlmsg_flags & NLM_F_ACK) || err)
-		rd_send_err(req, cps_conf, err);
-	return ret;
-}
-
-/*
- * Receive a netlink message with the ability to pass flags to recvmsg().
- * This function is an adaptation of mnl_socket_recvfrom() from
- * http://git.netfilter.org/libmnl/tree/src/socket.c#n263, which does
- * not allow flags.
- */
-static ssize_t
-mnl_socket_recvfrom_flags(const struct mnl_socket *nl, void *buf, size_t bufsiz,
-	int flags)
-{
-	ssize_t ret;
-	struct sockaddr_nl addr;
-	struct iovec iov = {
-		.iov_base = buf,
-		.iov_len  = bufsiz,
-	};
-	struct msghdr msg = {
-		.msg_name       = &addr,
-		.msg_namelen    = sizeof(struct sockaddr_nl),
-		.msg_iov        = &iov,
-		.msg_iovlen     = 1,
-		.msg_control    = NULL,
-		.msg_controllen = 0,
-		.msg_flags      = 0,
-	};
-	ret = recvmsg(mnl_socket_get_fd(nl), &msg, flags);
-	if (ret == -1)
-		return ret;
-
-	if (msg.msg_flags & MSG_TRUNC) {
-		errno = ENOSPC;
-		return -1;
-	}
-	if (msg.msg_namelen != sizeof(struct sockaddr_nl)) {
-		errno = EINVAL;
-		return -1;
-	}
-	return ret;
-}
-
-void
-kni_cps_rd_event(struct cps_config *cps_conf)
-{
-	unsigned int update_pkts = cps_conf->max_rt_update_pkts;
-	do {
-		char buf[MNL_SOCKET_BUFFER_SIZE];
-		int ret = mnl_socket_recvfrom_flags(cps_conf->rd_nl, buf,
-			sizeof(buf), MSG_DONTWAIT);
-		if (ret == -1) {
-			if (errno != EAGAIN && errno != EWOULDBLOCK)
-				CPS_LOG(ERR, "%s: recv: %s\n",
-					__func__, strerror(errno));
-			break;
-		}
-
-		ret = mnl_cb_run(buf, ret, 0, 0, rd_cb, cps_conf);
-		if (ret != MNL_CB_OK)
-			break;
-
-		update_pkts--;
-	} while (update_pkts > 0);
 }
 
 /*
